// src/src/model/requisition.ts
export interface IQuotationGrid {
  ID: string; //SharePoint Item Id
  PartNumber?: string; //Column "Part Number"
  Qualifier?: string; //Column "Qualifier"
  PartDescription?: string; //Column "Part Description"
  PartIssue?: string; //Column "Part Issue"
  DrawingNo?: string; //Column "Drawing No."
  Status?: string; //Column "Status"
  OrderType?: string; //Column "Order Type"
  MaterialUser?: number; //Column "Material User"
  Suffix?: string; //Column "Suffix"
  Porg?: string; //Column "Porg"
  HandlerId?: number; //Column "Handler"
  BuyerName?: string; //Column "Buyer Name"
  PARMA?: string; //Column "Parma"
  NamedPlace?: string; //Column "Named Place"
  NamedPlaceDescription?: string; //Column "Named Place Description"
  SurfaceTreatmentCode?: string; //Column "Surface Treatment Code"
  CountryOfOrigin?: string; //Column "Country of Origin"
  OrderCoverageTime?: number; //Column "Order Coverage Time"
  FirstLot?: string; //Column "First Lot"
  SupplierPartNumber?: string; //Column "Supplier Part Number"
  Currency?: string; //Column "Currency"
  UnitOfPrice?: string; //Column "UOP"
  UOM?: string; //Column "UOM"
  QuotedUnitPriceTtl?: string; //Column "Quoted Unit Price"
  OrderPriceStatusCode?: string; //Column "Price Type"
  QuotedToolingPriceTtl?: string; //Column "Quoted Tooling Price Ttl"
  QuotedOneTimePaymentTtl?: string; //Column "Quoted One Time Payment Ttl"
  MaterialsCostsTtl?: string; //Column "Materials Costs Ttl"
  PurchasedPartsCostsTtl?: string; //Column "Purchased Parts Costs Ttl"
  ProcessingCostsTtl?: string; //Column "Processing Costs Ttl"
  ToolingJigDeprCostTtl?: string; //Column "Tooling Jig Depr Costs Ttl"
  AdminExpProfit?: string; //Column "Admin Exp/Profit"
  PackingAndDistributionCosts?: string; //Column "Packing and Distribution Costs"
  Other?: string; //Column "Other"
  QuotedBasicUnitPriceTtl?: string; //Column "Quoted Basic Unit PriceTtl"
  PaidProvPartsCost?: string; //Column "Paid Prov Parts Cost"
  SuppliedMtrCost?: string; //Column "Supplied Mtr Cost"
  CommentHistory?: string; //Column "Comment History"
  AnnualQty?: string; //Column "Annual Qty"
  OrderQty?: string; //Column "Order Qty"
  RequiredWeek?: string; //Column "Required Week"
  RequisitionType?: string; //Column "Requisition Type"
}
export interface IProceedToPoFields {
  ID?: string;
  Status?: string;
  StandardOrderText1?: string;
  StandardOrderText2?: string;
  StandardOrderText3?: string;
  FreePartText?: string;
  CommentHistory?: string;
  OrderNumber?: string;
}